+++ common/common.tcpforward/TcpForwardTargetCaching.cs
@@ -10,7 +10,7 @@ namespace common.tcpforward
     public class TcpForwardTargetCaching : ITcpForwardTargetCaching<TcpForwardTargetCacheInfo>
     {
         private readonly ConcurrentDictionary<string, TcpForwardTargetCacheInfo> cacheHost = new();
-        private readonly ConcurrentDictionary<int, TcpForwardTargetCacheInfo> cache = new();
+        private readonly ConcurrentDictionary<ushort, TcpForwardTargetCacheInfo> cache = new();
 
         public TcpForwardTargetCaching()
         {
@@ -22,44 +22,44 @@ namespace common.tcpforward
             cacheHost.TryGetValue(host, out TcpForwardTargetCacheInfo cacheInfo);
             return cacheInfo;
         }
-        public TcpForwardTargetCacheInfo Get(string domain, int port)
+        public TcpForwardTargetCacheInfo Get(string domain, ushort port)
         {
             return Get(JoinHost(domain, port));
         }
-        public TcpForwardTargetCacheInfo Get(int port)
+        public TcpForwardTargetCacheInfo Get(ushort port)
         {
             cache.TryGetValue(port, out TcpForwardTargetCacheInfo cacheInfo);
             return cacheInfo;
         }
 
-        public bool Add(string domain, int port, TcpForwardTargetCacheInfo mdoel)
+        public bool Add(string domain, ushort port, TcpForwardTargetCacheInfo mdoel)
         {
             return cacheHost.TryAdd(JoinHost(domain, port), mdoel);
         }
-        public bool Add(int port, TcpForwardTargetCacheInfo mdoel)
+        public bool Add(ushort port, TcpForwardTargetCacheInfo mdoel)
         {
             bool res = cache.TryAdd(port, mdoel);
             return res;
         }
 
-        public void AddOrUpdate(string domain, int port, TcpForwardTargetCacheInfo mdoel)
+        public void AddOrUpdate(string domain, ushort port, TcpForwardTargetCacheInfo mdoel)
         {
             cacheHost.AddOrUpdate(JoinHost(domain, port), mdoel, (a, oldValue) => mdoel);
         }
-        public void AddOrUpdate(int port, TcpForwardTargetCacheInfo mdoel)
+        public void AddOrUpdate(ushort port, TcpForwardTargetCacheInfo mdoel)
         {
             cache.AddOrUpdate(port, mdoel, (a, oldValue) => mdoel);
         }
 
-        public bool Remove(string domain, int port)
+        public bool Remove(string domain, ushort port)
         {
             return cacheHost.TryRemove(JoinHost(domain, port), out _);
         }
-        public bool Remove(int port)
+        public bool Remove(ushort port)
         {
             return cache.TryRemove(port, out _);
         }
-        public IEnumerable<int> Remove(string targetName)
+        public IEnumerable<ushort> Remove(string targetName)
         {
             var keys = cache.Where(c => c.Value.Name == targetName).Select(c => c.Key);
             foreach (var key in keys)
@@ -74,16 +74,16 @@ namespace common.tcpforward
             return keys;
         }
 
-        public bool Contains(string domain, int port)
+        public bool Contains(string domain, ushort port)
         {
             return cacheHost.ContainsKey(JoinHost(domain, port));
         }
-        public bool Contains(int port)
+        public bool Contains(ushort port)
         {
             return cache.ContainsKey(port);
         }
 
-        private string JoinHost(string domain, int port)
+        private string JoinHost(string domain, ushort port)
         {
             return port == 80 || port == 443 ? domain : $"{domain}:{port}";
         }
