+++ server/plugins/server.service.tcpforward/TcpForwardTargetProvider.cs
@@ -22,7 +22,7 @@ namespace server.service.tcpforward
         {
             GetTarget(tcpForwardTargetCaching.Get(host), info);
         }
-        public void Get(int port, TcpForwardInfo info)
+        public void Get(ushort port, TcpForwardInfo info)
         {
             GetTarget(tcpForwardTargetCaching.Get(port), info);
         }
